backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
joblib
scikit-learn
pillow
xgboost
